tested and working api endpoint for getting, storing, and updating mentor preferences
