+++ __init__.py
@@ -9,11 +9,11 @@ from flask_mail import Mail
 from db import mysql
 from db_utils import *
 from pathlib import Path
-from resources.user import (UserRegister, Users, UserLogin, UserLogout, 
-							User, ResetPassword, CheckIfActive, UsersHighscores, 
-							UserLevels, GenerateUsername, GetUsernames, 
+from resources.user import (UserRegister, Users, UserLogin, UserLogout,
+							User, ResetPassword, CheckIfActive, UsersHighscores,
+							UserLevels, GenerateUsername, GetUsernames,
 							GenerateOTC, OTCLogin, User_Preferences,
-							ForgotPassword, ChangePassword, ForgotUsername, SaveMentorPreference)
+							ForgotPassword, ChangePassword, ForgotUsername)
 from resources.terms import (Term, Tags, Tag_Term, Tags_In_Term, 
 							 Specific_Term, TagCount)
 from resources.sessions import (Session, SearchSessions, End_Session, 
@@ -30,6 +30,7 @@ from resources.access import Access
 from resources.group import (Group, GroupRegister, SearchUserGroups, 
 							 UsersInGroup, GenerateGroupCode)
 from resources.logged_answer import LoggedAnswer, GetLoggedAnswerCSV
+from resources.mentors import MentorPreference
 import os.path
 import config
 
@@ -167,7 +168,7 @@ api.add_resource(TermsPerformance, API_ENDPOINT_PREFIX+'termsperformance')
 api.add_resource(ForgotPassword, API_ENDPOINT_PREFIX+'forgotpassword', resource_class_kwargs={'mail' : mail})
 api.add_resource(ChangePassword, API_ENDPOINT_PREFIX+'changepassword')
 api.add_resource(ForgotUsername, API_ENDPOINT_PREFIX+'forgotusername', resource_class_kwargs={'mail' : mail})
-api.add_resource(SaveMentorPreference, API_ENDPOINT_PREFIX + 'savementorpreference')
+api.add_resource(MentorPreference, API_ENDPOINT_PREFIX + 'mentorpreference')
 
 if __name__ == '__main__':
 	app.run(host='0.0.0.0', port='5050', debug=True)
