+++ utils.py
@@ -180,25 +180,20 @@ def check_user_db(_id):
 
     return False
 
-def check_mentor_preferences(_id):
+def get_mentor_preference(_id, conn, cursor):
+    query = "SELECT mentorName FROM mentor_preferences WHERE userID = %s"
+    result = getFromDB(query, _id)
 
-    query = "SELECT * FROM user WHERE userID = %s"
-    result = getFromDB(query, (_id))
+    return result
 
-    for row in result:
-        if row[0] == _id:
-            return True
+def store_mentor_preference(_id, mentor_name, conn, cursor):
 
-    return False
-
-def store_mentor_preferences(_id, mentor_name, conn, cursor):
-
-    if(check_mentor_preferences(_id)):
+    if(len(get_mentor_preference(_id, conn, cursor)) > 0):
         query = "UPDATE mentor_preferences SET mentorName = %s WHERE userID = %s"
         postToDB(query, (mentor_name, _id), conn, cursor)
         return True
     else:
-        query = "INSERT INTO mentor_preferences VALUES (%s, %s)"
+        query = "INSERT INTO mentor_preferences (`userID`, `mentorName`) VALUES (%s, %s)"
         postToDB(query, (_id, mentor_name), conn, cursor)
         return False
